Check CHANGELOG.md

+++ src/main/java/com/blocklegend001/immersiveores/item/ModArmorMaterials.java
@@ -27,7 +27,7 @@ public class ModArmorMaterials {
                 map.put(ArmorItem.Type.CHESTPLATE, VibraniumConfig.protectionValueVibraniumChestplate);
                 map.put(ArmorItem.Type.HELMET, VibraniumConfig.protectionValueVibraniumHelmet);
             }), VibraniumConfig.enchantmentValueVibraniumArmor, SoundEvents.ITEM_ARMOR_EQUIP_NETHERITE, () -> Ingredient.ofItems(ModItems.VIBRANIUM_INGOT),
-                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "vibranium"))), VibraniumConfig.toughnessValueVibraniumArmor, VibraniumConfig.knockbackResistanceValueVibraniumArmor));
+                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "vibranium"))), VibraniumConfig.toughnessValueVibraniumArmor, (float) VibraniumConfig.knockbackResistanceValueVibraniumArmor));
 
     public static final RegistryEntry<ArmorMaterial> VULPUS = registerArmorMaterial("vulpus",
             () -> new ArmorMaterial(Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
@@ -36,7 +36,7 @@ public class ModArmorMaterials {
                 map.put(ArmorItem.Type.CHESTPLATE, VulpusConfig.protectionValueVulpusChestplate);
                 map.put(ArmorItem.Type.HELMET, VulpusConfig.protectionValueVulpusHelmet);
             }), VulpusConfig.enchantmentValueVulpusArmor, SoundEvents.ITEM_ARMOR_EQUIP_NETHERITE, () -> Ingredient.ofItems(ModItems.VULPUS_INGOT),
-                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "vulpus"))), VulpusConfig.toughnessValueVulpusArmor, VulpusConfig.knockbackResistanceValueVulpusArmor));
+                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "vulpus"))), VulpusConfig.toughnessValueVulpusArmor, (float) VulpusConfig.knockbackResistanceValueVulpusArmor));
 
     public static final RegistryEntry<ArmorMaterial> ENDERIUM = registerArmorMaterial("enderium",
             () -> new ArmorMaterial(Util.make(new EnumMap<>(ArmorItem.Type.class), map -> {
@@ -45,7 +45,7 @@ public class ModArmorMaterials {
                 map.put(ArmorItem.Type.CHESTPLATE, EnderiumConfig.protectionValueEnderiumChestplate);
                 map.put(ArmorItem.Type.HELMET, EnderiumConfig.protectionValueEnderiumHelmet);
             }), EnderiumConfig.enchantmentValueEnderiumArmor, SoundEvents.ITEM_ARMOR_EQUIP_NETHERITE, () -> Ingredient.ofItems(ModItems.ENDERIUM_INGOT),
-                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "enderium"))), EnderiumConfig.toughnessValueEnderiumArmor, EnderiumConfig.knockbackResistanceValueEnderiumArmor));
+                    List.of(new ArmorMaterial.Layer(Identifier.of(ImmersiveOres.MOD_ID, "enderium"))), EnderiumConfig.toughnessValueEnderiumArmor, (float) EnderiumConfig.knockbackResistanceValueEnderiumArmor));
 
     public static RegistryEntry<ArmorMaterial> registerArmorMaterial(String name, Supplier<ArmorMaterial> material) {
         return Registry.registerReference(Registries.ARMOR_MATERIAL, Identifier.of(ImmersiveOres.MOD_ID, name), material.get());
